Title: 입력받아 계산, Time: 117ms, Memory: 24MB, Status: Wrong Answer - CodeTree

+++ 240122/입력받아 계산/input-calculate.py
@@ -1,3 +1,5 @@
 a = input() #문자열로 입력받음
-a = int(a)
+a = float(a)
 print(a + 2)
+
+#a = int(input())
